README.md: Explain how to clone the repository
When bootstrapping the Mercury repository on Windows, one has to ensure that Git does not automatically convert
line endings to CRLF on Windows. Not doing so will cause the configure.ac script to properly expand the runtime/*.in files and will fail the build.
The only way to properly prevent this is to modify the Git configuration before cloning the Mercury repository, thus this is now mentioned in README.md which is displayed by GitHub

README.md: see above